# .history/openapi/oapi_20190923142547.yaml
openapi: 3.0.0
info:
  title: CCCS
  version: "3.0.0"
servers:
  - url: http://0.0.0.0:8080/api/v3
    description: Local Development Server
  - url: https://api-staging.citizenscience.ch/api/v3
    description: Staging Server
  - url: https://api.citizenscience.ch/api/v3
    description: Production API Server
paths:
...
  /users:
    get:
      security:
        - oauth2:
            []
      tags:
        - Users
      operationId: get_users
      parameters:
        - name: search_term
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 0
            default: 20
      responses:
        "200":
          description: Return all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./schemas.yaml#/components/schemas/User"
      x-openapi-router-controller: api.user.model
...
  /users/register:
    post:
      tags:
        - Users
      summary: Registration end point for a user account
      operationId: create_user
      requestBody:
        $ref: "./bodies.yaml#/components/requestBodies/User"
      responses:
        "201":
          description: Created user
        "409":
          description: User exists
      x-openapi-router-controller: api.user.model
...
  "/users/{id}":
    get:
      security:
        - apiKeyHeader:
            []
        - anonUser:
            []
      tags:
        - Users
      summary: Get all users (or those matching an ID)
      operationId: get_user
      parameters:
        - description: The unique identifer for an Object (i.e. User, Task, Project,
            Submission etc)
          name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[a-zA-Z0-9-]+$
      responses:
        "200":
          description: Return users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./schemas.yaml#/components/schemas/User"
        "404":
          description: User not found
      x-openapi-router-controller: api.user.model
    put:
      security:
        - apiKeyHeader:
            []
        - anonUser:
            []
      tags:
        - Users
      summary: Update user details (change password, add details etc)
      operationId: update_user
      parameters:
        - description: The unique identifer for an Object (i.e. User, Task, Project,
            Submission etc)
          name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[a-zA-Z0-9-]+$
      requestBody:
        $ref: "./bodies.yaml#/components/requestBodies/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "./schemas.yaml#/components/schemas/User"
        "404":
          description: User not found
      x-openapi-router-controller: api.user.model
    delete:
      security:
        - apiKeyHeader:
            []
        - anonUser:
            []
      tags:
        - Users
      summary: Delete user (only allowed by the user themselves)
      operationId: delete_user
      parameters:
        - description: The unique identifer for an Object (i.e. User, Task, Project,
            Submission etc)
          name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[a-zA-Z0-9-]+$
      responses:
        "200":
          description: User deleted
        "401":
          description: Not authorised
        "404":
          description: User not found
      x-openapi-router-controller: api.user.model
...
components:
  securitySchemes:
    anonUser:
      type: apiKey
      name: X-ANON
      in: header
      x-apikeyInfoFunc: decorators.access_checks.ensure_anon_key
    apiKeyHeader:
      type: apiKey
      name: X-API-KEY
      in: header
      x-apikeyInfoFunc: decorators.access_checks.ensure_key
    oauth2:
      type: oauth2
      x-tokenInfoUrl: http://localhost:8081/api/v1/users/validate
      flows:
        implicit:
          authorizationUrl: http://localhost:8081/oauth/authorize
          scopes:
            apiKey: Unique identifier of the user accessing the service.